+++ src/app/models/disclaimer.model.ts
@@ -1,0 +1,8 @@
+export type DisclaimerFormStore = {
+  [key: number]: DisclaimerForm;
+};
+
+export type DisclaimerForm = {
+  understood: boolean;
+  experience?: string;
+};